ruby: break out ri-docs into a separate package
The ri (Ruby Interactive) documentation for the Ruby standard library
consumes a significant amount of space on disk.  It is useful to
developers, but is usually not necessary for users who just want to run
applications written in Ruby.  Break it out into a separate package so
Ruby can be installed without it.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Alejandro del Castillo <alejandro.del.castillo@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
---
Natinst-ReviewBoard-ID: 85087