php: fix build issue with shared native sstate
The PEAR build gets the location of its config file (pear.conf) from a
call into php-native.  Due to an improper use of absolute rather than
relative paths, pear.conf is erroneously generated under
${D}/${STAGING_DIR_NATIVE}.  In a do_install_append step for targets
only, the php recipe moves pear.conf to be under ${D}.  However, this
fails if the sstate cache for -native packages is shared between builds,
since pear.conf is always generated under ${STAGING_DIR_NATIVE} for the
target for which php-native was built, which will not match if we're
building for a different target.  This manifests itself as a build
failure for the second target to build php: on the feed build,
php-native and php successfully build for Zynq, but for x64, php-native
is reused from sstate cache but the php build fails.

To fix this issue, fix up the ${STAGING_DIR_NATIVE} path as appropriate
for either architecture we build for.

A cleaner fix would be to edit the installation scripts for PEAR, but
since these are contained in a .phar archive, patching the scripts
causes the checksum verification of the archive to fail, aborting the
build.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 87440