apache2: Add perl to apache2-dev RDEPENDS
RDEPENDS on perl in apache2-dev package because it ships perl scripts.

This is a cherry-pick of Armin Kuster's 1a3d7ed7 to apache2_2.4.12.bb
from meta-OE/master. We can drop this change once 1a3d7ed7 is included
in NILRT or the apache2 recipe is updated to 2.4.12 or higher.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-CAR-ID: 574179
Natinst-ReviewBoard-ID: 130047